python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "mastermatcher"
version = "0.1.0"
description = "Python backend for MasterMatcher VST plugin"
authors = ["Left Ear Audio"]

[tool.poetry.dependencies]
python = "^3.8"
matchering = "^3.0.0"
numpy = "^1.21.0"
scipy = "^1.7.0"
soundfile = "^0.10.3"
python-dotenv = "^0.19.0"
watchdog = "^2.1.0"
psutil = "^5.8.0"
pydub = "^0.25.1"
requests = "^2.26.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.flake8]
max-line-length = 88
ignore = "E501"
exclude = [
    "__pycache__",
    ".git",
    "dist",
    "build"
]
